fix: Some backend config issues are now resolved.
