+++ backend/src/main/java/com/boardwise/backend/shared/security/JwtFilter.java
@@ -1,38 +1,91 @@
 package com.boardwise.backend.shared.security;
 
 import java.io.IOException;
-import java.util.List;
+import java.util.HashMap;
+import java.util.Map;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
+import com.boardwise.backend.user_service.services.JWTService;
+import com.boardwise.backend.user_service.services.MyUserDetailsService;
+
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.MalformedJwtException;
+import io.jsonwebtoken.security.SignatureException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
+import tools.jackson.databind.ObjectMapper;
 
 @Component
-@RequiredArgsConstructor
-public class JwtFilter extends OncePerRequestFilter{
-    private final JwtUtil jwtUtil;
+public class JWTFilter extends OncePerRequestFilter{
+
+    @Autowired
+    private JWTService service;
+
+    @Autowired
+    private ApplicationContext context;
 
     @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException{
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
+            throws ServletException, IOException {
+        
         String authHeader = request.getHeader("Authorization");
+        String username = null;
+        String token = null;
 
-        if(authHeader != null && authHeader.startsWith("Bearer ")){
-            String token = authHeader.substring(7);
-            if(jwtUtil.validateToken(token)){
-                String username = jwtUtil.extractUsername(token);
-                UsernamePasswordAuthenticationToken auth = new UsernamePasswordAuthenticationToken(
-                    username, token, List.of());
-                SecurityContextHolder.getContext().setAuthentication(auth);
+        try{
+            if(authHeader != null && authHeader.startsWith("Bearer ")){
+                token = authHeader.split(" ")[1];
+                username = service.extractUsername(token);
+            }
+            
+            if(username != null && SecurityContextHolder.getContext().getAuthentication() == null){
+                UserDetails userDeets = context.getBean(MyUserDetailsService.class).loadUserByUsername(username);
+                
+                if(service.validateToken(token, userDeets)){
+                    UsernamePasswordAuthenticationToken authToken = 
+                    new UsernamePasswordAuthenticationToken(userDeets, null ,userDeets.getAuthorities());
+                    
+                    authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+                    SecurityContextHolder.getContext().setAuthentication(authToken);
+                }
             }
         }
-
+        catch (SignatureException e) {
+            handleJwtException(response, "JWT signature is invalid or expired");
+            return;
+        } catch (ExpiredJwtException e) {
+            handleJwtException(response, "JWT token has expired");
+            return;
+        } catch (MalformedJwtException e) {
+            handleJwtException(response, "JWT token is malformed");
+            return;
+        } catch (Exception e) {
+            handleJwtException(response, "Invalid JWT token");
+            return;
+        }
         filterChain.doFilter(request, response);
     }
+
+    private void handleJwtException(HttpServletResponse response, String message) throws IOException {
+        response.setStatus(HttpStatus.UNAUTHORIZED.value());
+        response.setContentType("application/json");
+        
+        Map<String, String> errorResponse = new HashMap<>();
+        errorResponse.put("message", message);
+        
+        ObjectMapper mapper = new ObjectMapper();
+        response.getWriter().write(mapper.writeValueAsString(errorResponse));
+    }
+
 }
