--- backend/src/main/java/com/boardwise/backend/user_service/config/SecurityConfig.java
@@ -1,63 +1,0 @@
-package com.boardwise.backend.user_service.config;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-import com.boardwise.backend.user_service.filters.JWTFilter;
-import com.boardwise.backend.user_service.services.MyUserDetailsService;
-
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    @Autowired
-    private MyUserDetailsService userDetailsService; 
-
-    @Autowired
-    private JWTFilter jwtFilter;
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http){
-        return http
-                .csrf(csrf -> csrf.disable())
-                .authorizeHttpRequests(request -> 
-                    request.requestMatchers("/api/auth/register", "/api/auth/login")
-                    .permitAll()
-                    .anyRequest()
-                    .authenticated()
-                )
-                .httpBasic(Customizer.withDefaults())
-                .sessionManagement(sess -> 
-                    sess.sessionCreationPolicy(
-                        SessionCreationPolicy.STATELESS
-                    )
-                )
-                .addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class)
-                .build();
-    }
-
-    @Bean
-    public AuthenticationProvider authenticationProvider(){
-        DaoAuthenticationProvider provider = new DaoAuthenticationProvider(userDetailsService);
-        provider.setPasswordEncoder(new BCryptPasswordEncoder(12));
-        return provider;
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration config){
-        return config.getAuthenticationManager();
-    }
-}
