+++ backend/src/main/resources/application.properties
@@ -12,6 +12,7 @@ r2.account-id=${R2_ACCOUNT_ID}
 r2.access-key=${R2_ACCESS_KEY}
 r2.secret-key=${R2_SECRET_KEY}
 r2.bucket-rulebooks=${R2_BUCKET_RULEBOOKS}
+R2_DEV_URL=${R2_DEV_URL}
 
 logging.level.org.springframework.boot.context.config=DEBUG
 spring.profiles.active=dev
